dash-practice/requirements.txt
dnspython==1.16.0
wheel==0.34.2
numpy==1.17.4
Flask-PyMongo==2.3.0
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
mkl==2019.0
pandas==0.25.3
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
retrying==1.3.3
WTForms==2.2.1
brotlipy==0.7.0
certifi==2020.4.5.1
cffi==1.14.0
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
Jinja2==2.11.2
MarkupSafe==1.1.1
plotly==4.7.1
pycparser==2.20
PyYAML==5.3.1
retrying==1.3.3
six==1.14.0
Werkzeug==1.0.1

